coach-agent_tmp20260127_051508/requirements.txt
google-cloud-aiplatform[agent_engines,adk]
google-adk
